Addressing Nikita's comment.
