+++ tests/platformalteration/platform_alteration_suite_test.go
@@ -38,6 +38,7 @@ var _ = BeforeSuite(func() {
 	err = globalhelper.DefineTnfConfig(
 		[]string{tsparams.PlatformAlterationNamespace},
 		[]string{tsparams.TestPodLabel},
+		[]string{},
 		[]string{})
 	Expect(err).ToNot(HaveOccurred())
 })
